Move Image Comment files into separate files
