+++ EssentialFeed/EssentialFeediOS/Image Comment UI/Controllers/ImageCommentCell.swift
@@ -1,0 +1,16 @@
+//
+//  ImageCommentCell.swift
+//  EssentialFeediOS
+//
+//  Created by Antonio Mayorga on 4/3/21.
+//  Copyright © 2021 Essential Developer. All rights reserved.
+//
+
+import Foundation
+import UIKit
+
+public class ImageCommentCell: UITableViewCell {
+	public let authorName = UILabel()
+	public let datePosted = UILabel()
+	public let comment = UILabel()
+}
