+++ EssentialFeed/EssentialFeediOSTests/Comment Presentation/ImageCommentViewControllerTest.swift
@@ -9,69 +9,7 @@
 import XCTest
 import UIKit
 import EssentialFeed
-
-final class ImageCommentViewController: UITableViewController {
-	private var loader: ImageCommentLoader?
-	private var errorView: UIView?
-	private var tableModel = [ImageComment]()
-	private var loadCommentTask: ImageCommentLoaderTask?
-	
-	convenience init(loader: ImageCommentLoader) {
-		self.init()
-		self.loader = loader
-		self.errorView = UIView()
-	}
-	
-	deinit {
-		loadCommentTask?.cancel()
-	}
-	
-	override func viewDidLoad() {
-		super.viewDidLoad()
-		
-		refreshControl = UIRefreshControl()
-		refreshControl?.beginRefreshing()
-		refreshControl?.addTarget(self, action: #selector(load), for: .valueChanged)
-		
-		load()
-		errorView?.isHidden = true
-	}
-	
-	@objc func load() {
-		loadCommentTask = loader?.load(completion: { [weak self] result in
-			switch result {
-			
-				case .failure(_):
-					self?.errorView?.isHidden = false
-					
-				case .success(let imageComments):
-					self?.tableModel = imageComments
-					self?.tableView.reloadData()
-			}
-			
-			self?.refreshControl?.endRefreshing()
-		})
-	}
-	
-	override func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-		return tableModel.count
-	}
-	
-	override func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-		let cellModel = tableModel[indexPath.row]
-		let cell = ImageCommentCell()
-		cell.authorName.text = cellModel.author.username
-		cell.comment.text = cellModel.message
-		cell.datePosted.text = cellModel.createdAt.description
-		return cell
-	}
-}
-
-class ImageCommentCell: UITableViewCell {
-	let authorName = UILabel()
-	let datePosted = UILabel()
-	let comment = UILabel()
-}
+import EssentialFeediOS
 
 class ImageCommentViewControllerTest: XCTestCase {
 	func test_loadCommentActions_requestCommentFromLoader() {
@@ -128,7 +66,7 @@ class ImageCommentViewControllerTest: XCTestCase {
 	}
 	
 	func test_viewDidLoad_hidesErrorMessage() {
-		let (sut, loader) = makeSUT()
+		let (sut, _) = makeSUT()
 		
 		sut.loadViewIfNeeded()
 		
@@ -214,76 +152,5 @@ class ImageCommentViewControllerTest: XCTestCase {
 										createdAt: date,
 										author: author)
 		return imageComment
-	}
-	
-	class LoaderSpy: ImageCommentLoader {
-		private var completions = [(LoadImageCommentResult) -> Void]()
-		private(set) var cancelledCompletions = [(LoadImageCommentResult) -> Void]()
-		var loadCallCount: Int { return completions.count }
-		
-		private class TaskSpy: ImageCommentLoaderTask {
-			let cancelCallback: () -> Void
-			
-			init(cancelCallback: @escaping () -> Void) {
-				self.cancelCallback = cancelCallback
-			}
-			
-			func cancel() {
-				cancelCallback()
-			}
-		}
-		
-		func load(completion: @escaping (LoadImageCommentResult) -> Void) -> ImageCommentLoaderTask {
-			completions.append(completion)
-			return TaskSpy { [weak self] in
-				self?.cancelledCompletions.append(completion)
-			}
-		}
-		
-		func completeCommentLoading(with imageComments: [ImageComment] = []) {
-			completions[0](.success(imageComments))
-		}
-		
-		func failedCommentLoading() {
-			completions[0](.failure(anyNSError()))
-		}
-	}
-}
-
-private extension ImageCommentViewController {
-	var isShowingLoadingIndicator: Bool {
-		return refreshControl?.isRefreshing == true
-	}
-	
-	var isShowingErrorView: Bool {
-		return errorView?.isHidden != true
-	}
-	
-	func numberOfRenderedImageCommentViews() -> Int {
-		return tableView.numberOfRows(inSection: imageCommentSections)
-	}
-	
-	func imageCommentView(at row: Int) -> UITableViewCell? {
-		let ds = tableView.dataSource
-		let index = IndexPath(row: row, section: imageCommentSections)
-		return ds?.tableView(tableView, cellForRowAt: index)
-	}
-	
-	private var imageCommentSections: Int {
-		return 0
-	}
-}
-
-private extension ImageCommentCell {
-	var authorNameText: String? {
-		return authorName.text
-	}
-	
-	var commentText: String? {
-		return comment.text
-	}
-	
-	var dateText: String? {
-		return datePosted.text
 	}
 }
